Enable GIT_SHALLOW
Previously, we were cloning the entire git history. Now we are only cloning the last commit. This should be a significant speed improvement for large repos and is equivalent to git clone --depth 1 <repo>. If you need the entire history later, git fetch --unshallow will pull the rest of the information.